+++ zhmcclient/__init__.py
@@ -32,6 +32,7 @@ from ._resource_updater import *       # noqa: F401
 from ._timestats import *     # noqa: F401
 from ._client import *        # noqa: F401
 from ._cpc import *           # noqa: F401
+from ._group import *         # noqa: F401
 from ._lpar import *          # noqa: F401
 from ._partition import *     # noqa: F401
 from ._activation_profile import *     # noqa: F401